+++ actionpack/CHANGELOG.md
@@ -1,3 +1,8 @@
+*   Fix `ActionDispatch::Executor` to unwrap exceptions like other error reporting middlewares.
+
+    *Jean Boussier*
+
+
 ## Rails 8.0.1 (December 13, 2024) ##
 
 *   Add `ActionDispatch::Request::Session#store` method to conform Rack spec.